+++ src/commands/query.ts
@@ -5,11 +5,10 @@ import spinner from "../services/spinner";
 import { authenticate, BaseOptions, baseOptions, printError } from "../shared";
 import { randomString } from "../utils";
 import handlers from "./query/handlers/handlers";
-import { promptServiceSelect as promptServiceSelect, promptFrom, promptQuerySelect, promptTo, promptGranularity } from "./query/prompts/query";
+import { promptFrom, promptQuerySelect, promptTo, promptGranularity } from "./query/prompts/query";
 import { getTimeframe } from "../services/timeframes/timeframes";
 
 export interface Options extends BaseOptions {
-  service?: string;
   id?: string;
   from?: string;
   granularity?: string;
@@ -23,7 +22,6 @@ export const builder: CommandBuilder<Options, Options> = (yargs) => {
   return yargs
     .options({
       ...baseOptions,
-      service: { type: "string", desc: "Name of the service" },
       id: { type: "string", desc: "Query id, if running a saved query" },
       from: { type: "string", desc: "UTC start time - may also be relative eg: 1h, 20mins" },
       to: { type: "string", desc: "UTC end time - may also be relative eg: 1h, 20mins, now" },
@@ -35,8 +33,8 @@ export const builder: CommandBuilder<Options, Options> = (yargs) => {
         # Run a query in interactive mode
         $0 query
 
-        # Run a saved query passing the service name and the queryId:
-        $0 query --service <service_name> --id <query_id> --from 2days --to 1day
+        # Run a saved query passing the queryId:
+        $0 query --id <query_id> --from 2days --to 1day
     `,
       ],
     ])
@@ -46,14 +44,12 @@ export const builder: CommandBuilder<Options, Options> = (yargs) => {
 };
 
 export async function handler(argv: Arguments<Options>) {
-  let { profile, id, from, to, granularity, format, service, "api-key": apiKey } = argv;
+  let { profile, id, from, to, granularity, format, "api-key": apiKey } = argv;
 
   spinner.init(!!argv.quiet);
   const config = await authenticate(profile, apiKey);
 
-  service ??= (await promptServiceSelect())?.name || "";
-
-  id ??= (await promptQuerySelect(service))?.id;
+  id ??= (await promptQuerySelect())?.id;
 
   let isSaved = true;
   if (!id) {
@@ -61,8 +57,8 @@ export async function handler(argv: Arguments<Options>) {
     id = `baselime-new-query-${randomString(6)}`;
   }
 
-  if (!(service && id)) {
-    throw new Error("Service and query id are required");
+  if (!id) {
+    throw new Error("Query id required");
   }
 
   from ??= await promptFrom();
@@ -74,7 +70,6 @@ export async function handler(argv: Arguments<Options>) {
   if (!isSaved) {
     return await handlers.interactive({
       queryId: id as string,
-      service: service as string,
       format,
       from,
       granularity,
@@ -88,7 +83,6 @@ export async function handler(argv: Arguments<Options>) {
     format,
     from,
     to,
-    service: service,
     granularity,
     config,
   });