+++ src/commands/report/handlers/common.ts
@@ -5,19 +5,19 @@ import { AlertCheck } from "../../../services/api/paths/alert-checks";
 import api from "../../../services/api/api";
 import outputs from "./outputs";
 
-export async function commonHandler(quiet: boolean, service: string, path?: string, format?: OutputFormat) {
+export async function commonHandler(quiet: boolean, path?: string, format?: OutputFormat) {
   spinner.init(quiet);
 
   const s = spinner.get();
 
   let status;
   if (path) {
-    status = JSON.parse((await readFile(path)).toString()) as { version: string; service: string; alertChecks: AlertCheck[] };
+    status = JSON.parse((await readFile(path)).toString()) as { version: string; alertChecks: AlertCheck[] };
   } else {
     s.start("Creating snapshots...");
-    const ids = (await api.alertsList(service)).map((alert) => alert.id);
+    const ids = (await api.alertsList()).map((alert) => alert.id);
     const promises = ids.map(async (id) => {
-      return await api.alertChecksCreate(service!, id, false, true);
+      return await api.alertChecksCreate(id, false, true);
     });
 
     const result = await Promise.all(promises);
@@ -28,7 +28,6 @@ export async function commonHandler(quiet: boolean, service: string, path?: stri
     status = {
       version: getVersion(),
       alertChecks: checks,
-      service,
     };
   }
   return status;