+++ src/commands/test/handlers.ts
@@ -4,12 +4,12 @@ import spinner from "../../services/spinner";
 import { getVersion, OutputFormat } from "../../shared";
 import outputs from "../report/handlers/outputs";
 
-async function test(format: OutputFormat, data: { service: string; outFile: string }) {
+async function test(format: OutputFormat, data: { outFile: string }) {
   const s = spinner.get();
   s.start("Checking alerts...");
-  const ids = (await api.alertsList(data.service)).map((alert) => alert.id);
+  const ids = (await api.alertsList()).map((alert) => alert.id);
   const promises = ids.map(async (id) => {
-    return await api.alertChecksCreate(data.service, id, false, true);
+    return await api.alertChecksCreate(id, false, true);
   });
 
   const result = await Promise.all(promises);
@@ -21,7 +21,6 @@ async function test(format: OutputFormat, data: { service: string; outFile: stri
     data.outFile,
     JSON.stringify({
       version: getVersion(),
-      service: data.service,
       alertChecks: checks,
     }),
   );