+++ src/services/api/paths/alerts.ts
@@ -27,7 +27,6 @@ export interface Alert {
   environmentId: string;
   userId: string;
   enabled: boolean;
-  service: string;
   channels: { type: ChannelTypes; targets: string[] }[];
   created?: string;
   updated?: string;
@@ -38,8 +37,8 @@ export interface Timeframe {
   to: number;
 }
 
-async function alertsList(service?: string): Promise<Alert[]> {
-  const res = (await client.get("/alerts", { params: { service } })).data;
+async function alertsList(): Promise<Alert[]> {
+  const res = (await client.get("/alerts")).data;
   return res.alerts;
 }