+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Deleted
+- The service field in alerts, queries and dashboards
 
 ## [0.0.81] 2023-10-13
 ### Deprecated